Two minor spec changes (#625)
* Remove mention of fixed reply channel name

Per the recent changes, the reply channel name is now at the discretion
of the server.

* Clarification on bytes/text values being empty

This commit clarifies that the bytes/text values need to be non-empty
(so not None/'') for them to be treated as present.

* HTTP Response headers are optional

During discussion with Andrew, we noted that they are optional, and that
that makes good sense. So this commit updates the spec to reflect
reality.